requirements.txt
colorama==0.4.6
pyperclip==1.9.0
python-decouple==3.8
python-tds==1.16.1
RapidFuzz==3.13.0